+++ src/pip/_internal/network/xmlrpc.py
@@ -1,6 +1,10 @@
+"""xmlrpclib.Transport implementation
+"""
 import logging
 
 from pip._vendor import requests
+# NOTE: XMLRPC Client is not annotated in typeshed as on 2017-07-17, which is
+#       why we ignore the type on this import
 from pip._vendor.six.moves import xmlrpc_client  # type: ignore
 from pip._vendor.six.moves.urllib import parse as urllib_parse
